[debug] Improve DBG() macro
DBG() is a hook which calls xDbgPrint() and then DbgPrint().  The macro
uses the ternary operator to accomplish the hook, and produces a code
path which can never yield false.  The compiler should optimize the
resulting code, thus saving any additional cost in space.  The hook is
thus accomplished for free.  It's a hook in the sense that args from DBG()
are passed through to DbgPrint(), but not before we inject another
function call with unrelated and position-dependent args.