[registry] Added Michael Brown's registry routines from sanbootconf
Michael Brown's registry manipulation functions in his "sanbootconf"
package for helping MS iSCSI drivers to work with gPXE (especially
for Windows XP, where you need a static IP address) are really quite
handy.  This patch adds them, and starts to make use of them.

Unfortunately, I forgot to release a previous bug-fix for a fairly
serious bug, so this commit is functional, but not the end of changes
to registry.c.  After upping the version and releasing, changes to
registry.c will continue.