[driver] Remove bus and disk details from the device extension
The eventual goal is for each module to know what it needs to know
and to delegate module-specific tasks to the appropriate modules.

In working towards this goal, the device extension in driver.h no
longer contains space for bus or disk structures.  This involved
major access changes.  Where modules used to access the bus/disk
members of the device extension, they now use helper functions to
get a pointer to these structures.

The helper functions get_bus_ptr() and get_disk_ptr() are in the bus
module.  This isn't really right, since the bus module should
eventually know nothing about disks.  It's a temporary measure.

The bus module has always been responsible for adding child devices
as well as the bus device itself.  When adding a device, the bus
module now allocates enough device extension space to include the
driver-common structure as well as the appropriate bus- or disk-
specific structure.

It might be good if other modules inform
the bus module of how much device extension space they need in the
future.  This commit does not introduce such logic, but it's an
idea for later.