[bus] Fix AoE unmounting
When we invalidate our bus relations, we are queried for them.  Each
of the PDOs that were previously known are dereferenced, so we need to
balance that by referencing each PDO we report before the report is
complete.

This is the fix for item 10 in V.'s readme.txt (I believe).