[irp,driver] Add mini IRP handling stacks
An irp__handling structure esentially matches an irp__handler
function to the IRP major and minor functions it is intended to
handle.  Booleans are used for the case where such a handler handles
multiple major and/or minor combinations.

In order to facilitate knowledge of when an Irp has been fully
handled for the driver module, we implement a boolean completion
status parameter.  If an irp__handler wants to prevent the Irp
from further processing by the driver Irp dispatch routine, it
should set the completion status to TRUE.

This miniature IRP handling strategy is sitting doing nothing until
routines are ported to use it.