[fixups] Clean compilation
Notably:

Commit f92581dcd0d3e80f5a4b26959a3a5621d8378c83 should not have
changed ULONG_PTR to winvblock__uint32_ptr.  ULONG_PTR is not
an unsigned long *, but an unsigned long which pointers are
sometimes cast to and from.

ExAllocatePool() is deprecated and so has been wrapped with a
macro which instead calls ExAllocatePoolWithTag() with the tag
'klBV' ("VBlk").

NdisQueryBuffer -> NdisQueryBufferSafe
NdisBufferVirtualAddress -> MmGetSystemAddressForMdlSafe

Some cast fix-ups.

Cleanly compiled under a variety of build environments:
2000 x86
XP x86
2003 x86
2003 x64/IA-64
Vista x86