[driver] Use pre-installed bus if possible
To ensure that we function, we've historically forced the creation
of the bus device.  The documentation suggests that users use the
Add New Hardware Wizard to add the driver.  This resulted in two bus
devices: The proper one known as a SCSI adapter, and a boot-up
unknown device which was annoying.  For PE environments where our
driver is very simply injected through TXTSETUP.SIF entries, the
boot-up unknown device bus will still be generated.