[driver] OsLoadOptions support needed for Recovery Console
We now make a note of the OsLoadOptions during registry checking
and implement get_opt() in the driver module for finding a
specific WinVBlock option.  WinVBlock options should appear in
BOOT.INI or TXTSETUP.SIF as a switch like this:

/WINVBLOCK=option1=value1,option2=value2

Recovery Console requires our disks to be discovered before the
driver re-initialization routine.  Using the above feature, we
will now execute the re-initialization routine early if the user
requests, since this routine will create a bus object.  The user
requests this with an OsLoadOption like this:

/WINVBLOCK=BUS=1