[disk] Split disk IO into AoE and RAM disk modules
The disk structure now includes a function pointer member to
represent the appropriate I/O routine for the different disk
classes.  These classes are currently AoE and RAM, but could
include others in the future, such as file-backed disks.

disk__io() is a function which calls the appropriate routine
for a given disk (via the device extension space for now).

The AoE module no longer has any knowledge of RAM disks.  Yay.