[disk_pnp] Split ID query logic for disk class
The PnP ID query now calls disk__query_id() which in turn
calls the disk class-specific function to return the right
ID(s) in the passed wide-char string buffer.

There should no longer be any checks anywhere for the
IsRamdisk boolean member of the disk structure.  Now we
are free to remove that member and re-arrange how we use the
device extension space, rather than having a union for our
difference disk classes.