[cosmetics] Run indent twice due to indent bug
You might or might not have noticed that 'indent' will take
a file and change its mind about how to indent it.  Take that
resulting file and run 'indent' again and it will put the file
back the way it was.  Thus there is an "even" format and an
"odd" format.  This is silly.  As a work-around, the indent.sh
and indent_add.sh scripts will run 'indent' twice on a file.

Yeesh.