[driver,bus] Remove re-initialization routine
MAJOR LEARNING: A boot driver re-initialization routine is
too late to create the bus device and its disks!  Why?
If the boot volume was not installed during a previous
session, that means it's a new volume.  Never-before-seen
volumes are only registered as possible boot volumes at a
point before boot driver re-initialization routines are
called.

As such, our previous re-initialization routine was _only_
useful if the expected boot volume on one of the RAM disks
was somehow already installed...  It so happens that in
testing, I often modify the RAM disk image by using it
with QEmu as the virtual machine's hard disk drive.  Thus,
the volume gets installed at that time and persists.  The
Blue Screen of Death that folks were having was pretty
mysterious.  The symptoms involved the Session Manager
crashing at a post-0x0000007B time, which was unexpected.

The crash was due to an uninstalled volume being installed
too late, as mentioned.  So now probe__disks() happens
at bus creation time, and bus creation time happens either
during DriverEntry (if the user requests) or when the PnP
Manager re-establishes a previously installed bus device.
Both of these times occur before a boot driver
re-initialization routine occurs.

An easy way to reproduce the issue was to change the MBR
signature on a disk already installed, thus making it a
"brand new" disk, with "brand new" volumes on it.

This also explains why the Recovery Console will only boot
from a WinVBlock RAM disk with the /WINVBLOCK=BUS=1 switch.
This switch caused early bus creation _and_ early disk
probing.

EXTRA NOTE: Including the previous commit in CHANGES.log,
which was missed.