[bus] Remove disk details from Bus_AddChild
A fair bit of shuffling.  Devices extensions now have
a device operations member which is a struct of function
pointers; one for device creation and one for device
initialization.  This is akin to the disk operations
structure found as a member of a disk.

Bus_AddChild() is now simply passed a device extension
pointer, which points at device extension details to add
to the bus.  It will call the particular PDO creation
routine provided by whoever populated the device
extension.  That routine will return a PDO that will be
attached to the bus.  Bus_AddChild() is also still
passed a pointer to the bus device object to which the
child belongs.

The only PDO creation routine at this time is obviously
disk__create_pdo(), since we are a virtual SCSI driver!
The reason for this bit of abstraction is in case anyone
finds the code handy some day as a reference for the
bus-and-child model.  This code is not at that stage yet.