[build] Move majority of code into bus subdir
The plan is to split AoE functionality off into its own driver,
to avoid WinVBlock's dependency on NDIS for those users who
do not require or desire the AoE feature.

Similarly, future features with dependencies can have their own
directories and their own .SYS drivers.

For now, all driver code including AoE has been moved into bus/,
but AoE will eventually come out of there.

A strategy needs to be determined to allow for the single bus
device to accept features in a modular fashion, and to route
applicable feature-wide control signals to each feature's driver.

For example, the WinVBlock bus (an "HDD controller" of sorts) can
support certain virtual disk types natively.  When an AoE driver
starts, it could look for the bus and register itself as a
feature if the bus is found.  Thenceforth, a user-land utility
could request an AoE-attach of the WinVBlock bus, who could forward
that request on to the AoE driver.