[irp] Move mini IRP handling into its own module
We introduce bus/irp.c for the miniature IRP handling logic.

We also introduce code for an IRP handling table chain.  The
goal of such a chain is to allow IRP handlings to be
dynamically added and removed for a device.  An example use
would be for the AoE system to register IOCTL handling with
the WinVBlock bus device, so that the bus processes the
IOCTLs.  If the AoE system is unloaded, it should be able to
un-register the handling of these IOCTLs.

Currently, the IRP handling table chain is not used, but
this will change in a future commit.