[irp] Use mini IRP handling chain
When an IRP dispatch comes in, we special-case a power
IRP, or pass the IRP along to irp__process().

irp__process() will walk each table of IRP handlings
in the device's chain of registered tables, looking for
an IRP handling that matches the IRP major and minor
codes.  This is roughly similar to the PDO stacked
beneath multiple FDOs approach, but just a little less
complex and with perhaps less memory and management
overhead.

A feature (provided by a driver) can register handled
IRP types on a device, then WinVBlock should use that
feature's functions to process the IRP.  I hope WinVBlock
can call functions in another driver in this way. :)

So now when a bus or disk device is created, the default
driver IRP handling table is registered as well as the
device-specific handling table (bus or disk).  We used
to manually build a large table with these tables copied
into it, but not anymore.