[bus,aoe,irp] AoE mini IRP handling in AoE module
The bus module pretty well needs to expose the bus device to
other drivers so that they can register their mini IRP
handling tables with it.  bus__dev() now yields the FDO.

The un-register function was mis-named in the IRP module.

We are now registering the AoE IOCTL handling twice, but
this is not a big deal.  When AoE is finally split off,
the bus module will no longer register anything AoE-related.