[bus] Create boot bus here, introduce init function
The bus module is now responsible for creation of bus PDOs,
such as the boot bus which exposes WinVBlock-supported boot
disks.  We now have bus__init() and bus__finalize() funcs
which are called by the driver module.

We also mimic the device module by having a bus__create()
function, though it is currently unused, also like the
device__create() function.