[device,bus] Further develop device routines
Much of it unused for now, but eventually (seems like it's
always eventually) each device type will create an instance
of a type it would like to derive from, then overrides
whatever operations it wants to.  When the device is asked
to perform operations, it should either have sensible, do-
nothing defaults or meaningful device-specific operations.