[project] Change RtlCompareMemory to wv_memcmpeq
We previously had patterns like:

  if (!(RtlCompareMemory(foo, bar, sizeof foo) == sizeof foo))

but now we have:

  if (!wv_memcmpeq(foo, bar, sizeof foo))

We've introduced a few new headers, which have the spirit of
mimicry of the C Standard Library....

'memcmpeq()' is completely bogus, however.

The MS compiler doesn't have '_Bool', so we use unsigned short
for the 'bool' type.  I hope that's good enough.