[device] Make room for threaded devices
After thinking about a "handler" scheme for devices which
implement various "interfaces," it looked like it'd be
better to implement threading for devices first and to find
out if that would do.

We make space for queuing IRPs and for a thread wakeup signal.

We also replace a lonely ExFreePool() that was missed in a
previous commit with wv_free().