[irp] Move code into irp__process_with_table
We only have 4 mini IRP handling tables, so the
register/unregister functionality hardly seems worth it.  One
of the original reasons why it was added was so that AoE could
register with the bus device to receive IOCTLs.  We could do
it differently:

1. AoE could implement its own bus device as a sub-bus of the
   WinVBlock bus.

2. AoE could hook the WinVBlock bus' mini IRP handling table.
   This might be dangerous if AoE didn't unload properly.

3. AoE could still register with the WinVBlock bus to receive
   IRPs, but as a bus registration function versus an abstract
   IRP registration.

This commit paves the way for 'driver', 'bus', 'disk', 'aoe' to
each have their own IRP handling strategy, but to provide the
convenience of still using the mini IRP handling table strategy.
It removes emphasis from the mini IRP handling table
registration functions.