[bus,device] Allow for threaded PDO creation on a bus
bus__add_child() will add a PDO creation request to the
work item queue for a threaded bus, then wait for it to
complete.  We do this so that the PDO is created in the
bus' thread context.  For example, if the AoE driver were
to request a PDO, we want that PDO created by the WinVBlock
driver (the main bus' thread should be that context).

We add a completion signal pointer to work items, initialize
the signal upon adding a work item to a bus' queue, and
signal completion when a work item has been completed.