[bus] Introduce PDO node addition/removal functions
WV_S_BUS_NODE is an opaque type representing a child
PDO node on a bus.

WvBusInitNode() will initialize such a node and associate
it with the DEVICE_OBJECT that is its physical device
object.

WvBusAddNode() will schedule the addition of a node to a
bus.  When bus__process_work_items() is called for the
bus, the addition will take place.

WvBusRemoveNode() will schedule the removal of a node
from a bus.  When bus__process_work_items() is called
for the bus, the removal will take place.

bus__process_work_items() should be called by the bus'
thread routine, because it expects to have sole control
over the linked list of bus nodes (no race conditions,
please).

And of course, you might notice the i'th coding style
change.  This one is more of a "when in Rome..." change.

  WV_ttt_mmm_nnn

  WV  : "WinVBlock"
  ttt : type; can be a combination of
      E : enum
      S : struct
      U : union
      A : array
      F : function
      P : pointer
      I : integer
      M : #defined macro
  mmm : module
  nnn : type name