[bus] Remove members dev_name and dos_dev_name
Move responsibility for bus device naming out of the bus
module and into the caller.

The AoE module is a bit mixed-up just now and doesn't
work.  It's currently going to try to create a symlink
but the PDO for the AoE bus is not going to have been
named by the bus module.  What we really want is for the
AoE bus' _FDO_ to be the named device...  I think.