[bus] Introduce WvBusEnqueueIrp function
For a threaded bus, it is handy to keep all operations
that walk the list of child devices or use their count
in the thread to avoid race conditions and holding locks.
A IRP such as IRP_MN_QUERY_DEVICE_RELATIONS needs to
allocate a response whose size depends on the count of
child devices, for instance.  Such an IRP would be nice
to enqueue for the thread.