[bus] Add custom work items to WvBusProcessWorkItems
To be honest, a bus probably shouldn't be particularly
concerned with providing a default thread; it should be
up to the driver using the bus to ensure a single,
controlling thread for those operations which deal with
subjects such as the list of child nodes on the bus.

But for now, since there's already a mechanism in place
to enqueue work items for execution in a bus'
controlling thread, we abuse it to allow a driver to
enqueue their own driver-specific work item.  Use the
new WvBusEnqueueCustomWorkItem() function and the new
WV_S_BUS_CUSTOM_WORK_ITEM type.

This might change at some point, such as if thread
subject matter is removed from the bus library.