[driver] Introduce the dummy device feature
A driver can use WvDriverAddDummy() to produce a PDO on
the main WinVBlock bus.  The key point is that the PDO
will respond to PnP ID queries using a handler provided
by whichever driver calls WvDriverAddDummy().  This allows
for a driver to produce PDOs with arbitrary PnP IDs, which
is perhaps a bit better than IoReportDetectedDevice() for
our purposes.  The AoE driver could produce a bus PDO
using this technique, then attach an FDO to it.