[bus] Change #include guard
It's quite possible that I'll never settle on a naming
convention.

WVL_* and Wvl* for WinVBlock library stuff.  Then we
can remove the underscore suffix convention, which has
so far been in place to denote functions and objects
with translation unit-internal linkage.  Since these
outnumber those which are library exports, it might
make more sense to denote library exports.