[bus,dummy,libbus] Handle dummy PDO removal
The bus library can unlink a node, but should not be
responsible for deleting the corresponding PDO nor
for freeing associated resources.  A driver can test
for removal by checking the Linked member of the
WVL_S_BUS_NODE that is associated with the device.

This allows for the AoE bus to be uninstalled from
Device Manager without crashing!  What was happening
was that:
- The AoE bus got IRP_MN_REMOVE_DEVICE
  - It passed it on to the PDO
    - The dummy PDO handler unlinked the node and
      deleted the device.
    - It also triggered a re-enumeration of the
      WinVBlock bus.
- AoE shutdown
- The re-enumeration of the WinVBlock bus would find
  the PDO missing and send IRP_MN_REMOVE_DEVICE.
- The device was already deleted!

So now we merely unlink on the first removal IRP,
then delete the PDO on the second IRP, which will be
triggered by bus re-enumeration.