[debug] Change macros and message layout
This seems simpler.  We used to have #ifs for verbose IRP
debugging.  That' no longer needed.

DBG() could get pretty messy, so now we can choose to
exclude the file, line number, or thread.  Including the
line number depends on including the file.

Also, we exclude the bulk of debug.c from being compiled
if debugging is disabled.