[filedisk] Make all filedisks threaded
We don't have any control over how long it will take
for file I/O to complete (network share, for example),
so we will pend all SCSI reads/writes and each filedisk's
thread will complete them.

Hot-swapping a sector-mapped disk to a file-backed disk
for any number of disks is done from a single, special
thread.  Since GRUB4DOS is the only case where such info
is passed, GRUB4DOS is the only caller for the function
which starts up this special thread.  It's really quite
messy right now and needs a major clean-up.

The reason why the hot-swapping thread is its own thread
is to avoid the dead-lock where NTFS is trying to mount
a partition, then if the same thread tries to open a
file (to test for an appropriate file-backing) on an
NTFS volume, NTFS is already "busy".

The unfortunate thing is that the hot-swapping thread
doesn't guarantee a file lock as early as possible.
This can be improved, some day.