[filedisk] Add impersonation support for network files
It should now be possible to attach a file-backed disk to
a file on a network share, assuming the share has been
mapped to a drive letter and assuming permissions allow.

A to-do item would be to allow attaching read-only media
(such as optical discs) without trying for write
permissions on the file.

As a bonus, this commit means that user "User" can't
open files belonging to user "Administrator".  That is
a big security win and something that was pretty bad
before now.

Some re-ordering of functions and function comments for
other filedisk functions were added, too.