[filedisk/grub4dos] Pend WvFilediskG4dIo_
If we attempt to perform ZwReadFile() in a thread
context where there is already a read happening
(such as when Windows' SETUPDD is attempting to
mount our disk), then we will deadlock.  So we
pend the IRP and execute it within our filedisk's
thread's context.  This probably has to do with IRQL.