[device,dummy,bus,disk] Switch over to new IRP handling scheme
All WinVBlock driver devices, whether bus FDO, a dummy PDO, a
filedisk PDO, or a RAM disk PDO, all have a common, initial
WV_S_DEV_EXT structure in their DeviceExtension area.  This
includes a function pointer which points to the IRP dispatcher
function for the "class" of device.  So we simply pass the IRP
on to that routine.

Also fixed a bug in the dummy IRP handling where the device
extension was being passed instead of a device, itself.