[driver] Re-introduce re-initialization routine
After some more learning about the use of re-initialization
routines, it's time to put one back in.

Users wish to boot from a GRUB4DOS sector-mapped disk on a
USB backing disk.  I coded the "WaitBt driver" to assist
with the case of booting from a USB disk, but this is not
quite the same...  What we want is to avoid reporting disks
that have not yet arrived and thus avoid the "dud" disks
that we get when a backing disk cannot be found.

Instead, we ought to search for the backing disk first,
then report the disk once its backing disk has been found.
Since this might occur after all boot drivers have been
initialized, we need a re-initialization routine.

I _think_ that for USB disks, the Microsoft USB hub driver
can report no children at first, but then a poll might
find children (USB disks) later on.  My guess is that the
poll would be asynchronous to the boot device establishment
thread and result in crossed fingers that the poll and
USB disks happen before boot device establishment is
considered to be complete.

In addition!: If the USB disk arrives after the WinVBlock
disk, then the current bug where WinVBlock will search its
own disk would prevent us from ever finding the USB disk
as the backing disk.

So it seems that what we need to do is to delay until all
backing disks are found, or until some maximum number of
attempts has been reached; whichever comes first.