[filedisk] Attempt to stall driver reinitialization
This is a failed attempt to provide GRUB4DOS file-disks
from a reinitialization routine.  It didn't work.

I guess I'll try registering for disk arrival notifications
and examining disks as they arrive.  The only "problem" with
this would be that WinVBlock would have to register for these
very early, so as not to miss any disk arrivals.