[grub4dos] Introduce GRUB4DOS bus
Not yet finished, but the idea is that when WinVBlock starts,
it'll walk the INT 0x13 "safe hook" chain and make a "bus" for
each such hook.  Well actually, it'll make a bus for the first
hook and that bus will have a sub-bus if it finds another hook.

So, for example, if one has a GRUB4DOS INT 0x13 hook with all
drive mapping slots populated, there should be a GRUB4DOS bus
with 8 child disks.  If another "safe hook" in the chain is found,
there would also be a child bus.  A GRUB4DOS bus thus can have a
maximum of 9 child devices.

Since the WinVBlock driver will handle both the PDO and FDO
aspects of the GRUB4DOS bus, the AddDevice should be a null-op.

There should be a MEMDISK bus, too, at some point.