[safehook] Introduce safe INT 0x13 hook bus
A "safe INT 0x13" hook will be a bus device.  If there
is another hook in the chain, that will be a child
device.  For a particular safe hook vendor (such as
MEMDISK or GRUB4DOS), we could attach an FDO for
additional functionality, such as exposing disks as
child devices.