[safehook] Bus PDO probing, creation, walking
The WinVBlock bus will perform a probe of the INT 0x13 vector.
If a "safe hook" is found, a bus PDO for that hook will be
created.  When that bus is queried for its children, it will
perform a one-time check of the next hook in the chain and
determine whether or not there is a child bus for that hook,
and so on.

Things are temporarily broken until the GRUB4DOS and MEMDISK
buses are implemented as FDOs on top of a safe hook PDO.