[bus] Allow installation of raw PDOs
"Safe INT 0x13 hook" bus devices are driven entirely by the
WinVBlock driver, so allow them to be installed with the .INF
file and allow for the AddDevice routine to succeed if the
PDO is already driven by WinVBlock.

A side effect to note is that some other WinVBlock PDO (such
as a disk) could be forced to install by a user...  That
is not particularly useful to the user, however.