[driver] Introduce device threads
Two new functions that mini-drivers can use: WvlCreateDevice and
WvlDeleteDevice, which parallel IoCreateDevice and IoDeleteDevice.

When a device is created with WvlCreateDevice, it will have a
thread started for processing the device's IRPs and other work
items that need to be completed in a context in which the majority
of the device state is race-free.

This means that all of a device's IRPs will be processed
synchronously.  There is a performance penalty associated with
this, but until such a day as there's something better, this
strategy will do.  Right now, the focus is on functionality and
bug-fixes and to-dos and features.  Performance can come later.