[mainbus] Initial shuffle to mini-driver style
Removed WvBusCleanup.  Since could be called from an unload routine,
it shouldn't remove a device object, since those should already be gone.

WvDriverDispatchIrp will process the mini-driver case as well as the old
logic.

Right now, since some PnP IRPs cannot return STATUS_PENDING, all IRPs are
waited on for the mini-driver case.  WvProcessDeviceIrp was adjusted to
return information to WvDriverDispatchIrp.  Unfortunately, this somewhat
defeats the purpose of having a thread for every device, but not
completely.  A SynchronizationEvent might be the final solution.

In WvlCreateDevice, the caller wasn't receiving a pointer to the
newly-created device.  Heh.

Changed WvlDeleteDevice to simply signal a thread that it's time to
delete the device.  Now WvDeviceThread is solely responsible for deleting
a mini-driver device.  WvStopDeviceThread was adjusted to avoid requiring
the device after the thread has stopped.

In WvlCallFunctionInDeviceThread, the caller's function wasn't being used
in the work item.  Heh.

libbus/pnp was hacked to special-case the main bus' removal, since the
main bus is the only module on its way to being the mini-driver style.

Removed WvBusAttach and replaced with WvMainBusDriveDevice and
WvMainBusDriveDeviceInThread, instead.

The main bus' symbolic link is now cleaned up in either WvBusEstablish or
WvMainBusRemove.

WvMainBusOldProbe will need to go away once the other modules are
"internal mini-drivers."

The initial "floating FDO" strategy will probably go away, soon.